Try to detect intel-nvidia optimus configuration
After we install the system we try to detect an intel VGA/3D controller
and a Nvidia one. If we found that we ask the user if it likes to use
it. Answering NO will just do nothing. Answering YES will set up a
layout configuration using modesetting driver as primary GPU and nvidia
driver as secondary driver. Then the user can use nvidia-prime or set
the aproppriate variables to run games offloading frames on the
dedicated GPU.

This setup Nvidia claims to be also supported with AMD iGPUs and both
intel and amdgpu Xorg drivers. Those configurations/combinations should
be tested but maybe just using modesetting is the safest bet (and
detection of AMD iGPU should be added).

Note that AMD configurations just need PRIME=1 to be set to offload
frames and current X11 autodetection should work on those setups.